Generically handle auxiliary state
There seems to be a trifecta here:
1. dom nodes
2. custom elements
3. lists

From these three primitives you can build anything. The idea is to have
mount() only support those three element types and have html() generate
those primitives. For this to work, we need to be able to pass auxiliary
state to slightly tune these three primitives for html().

Hopefully we'll be able to get rid of the auxiliary state in the future,
but it's better than the special casing we were doing before.