transform/unsafeVariable: Implement live analysis
The previous algorithm would demote all variables to be vars. This is
really dangerous ergo the name of the pass. This new algorithm does
not denote all the variables to vars anymore, it instead does live analysis
of the variables consistent with the lifetimes given by the let or const.

This is safer and also easily lets us implement a rudimentary variable
reuse algorithm as well to allocate less variables. Cuts the bundle
by ~30 bytes

Note that this pass is still unsafe: variables that are captured by
closures won't have the same semantics as let or const.