dom: Remove broken optimization
This was originally meant to prevent moving around a node around on its
parent in case it was already mounted through static analysis. This
breaks real uses cases (the "static reuse" test tests for this).

Browsers are slow when laying out the dom. It doesn't matter as much if
the component is shifted momentarily while things are being built before
layout. Let's just remove the questionable optimization.